refactor: split LottoServiceCore into Base/Desktop/Mobile hierarchy
Replace single LottoServiceCore with a 3-tier class hierarchy:
- LottoServiceBase (abstract): shared constructor, check, getCheckWinningLink
- LottoServiceDesktop: desktop browser flow (Puppeteer/Playwright)
- LottoServiceMobile: mobile web flow (TODO stubs with LottoError.NotSupported)

Also fix LottoServiceInterface: purchase optional param, check param
name (volume → round), signInWithCookie param name (cookie → cookies).
