+++ src/index.ts
@@ -1,4 +1,4 @@
-export { LottoService } from './services/lottoService';
+export { LottoServiceDesktop, LottoService } from './services/lottoService';
 export { LogLevel } from './logger';
 export * from './types';
 export { getLastLottoRound } from './utils/getLastLottoRound';
